Moved serial binding to System.IO.Bindings.
