+++ ElfConsoleApplication/MBusEventSubscriber.cs
@@ -1,5 +1,4 @@
-﻿using MeterBusLibrary;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
