+++ ElfConsoleApplication/Settings.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO.Bindings.Serial;
 using System.Linq;
 using System.Text;
 
 namespace ElfConsoleApplication
 {
-    public sealed class Settings : MeterBusLibrary.SettingsSerial
+    public sealed class Settings : SettingsSerial
     {
         public override string PortName { get { return "COM5"; } }
