+++ System.IO.Bindings/Serial/Settings.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace MeterBusLibrary
+namespace System.IO.Bindings.Serial
 {
     public abstract class Settings
     {
